Move Kantele.Application into ExVenture.Application
Also host the websocket listener in the phoenix endpoint
